Release of V0.3: Core Boost
  Core: H4CK3D is substaintially bigger, but more durable and can hold
more items.

  FL4GSH1P now deals more damage (almost double it's damage).

  V3CT0R's Description no longer states that it is shielded. (Since that
proves it to be durable too, and V3CT0R is too deadly)

  W4RH0UND's Repair goes down from 6k to 3.5k.

  FL4GSH1P's Description got a rework, not stating any abilities.
(Doesn't need any abilities, loads of health and damage)

  Core: H4CK3D's Description got a rework, stating it's new stats, and
gives a notice of other mod's colliding with it. (I attempt to work
along side Exotic Mod but other mods that add lots of content may/will
collide with the main Core)